Converted SessionController into a singleton and squashed its public API
